requirements.txt
discord.py==1.4.1
google-auth==1.20.1
google-auth-oauthlib==0.4.1
gspread==3.6.0
jira==2.0.0
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.1.0
python-chess==0.31.4
python-dotenv==0.14.0
svglib~=1.0.0
reportlab~=3.5.47